fix warning

+++ src/assemble.cpp
@@ -193,7 +193,7 @@ void pamToPng(std::string png_name) {
   png_byte* pngRow = new png_byte[g_width * 4];
   for (int32_t i = 0; i < g_height; ++i) {
     if (pam) {
-      fread(pngRow, 4, g_width, pam);
+      fread(pngRow, 4, static_cast<size_t>(g_width), pam);
     } else {
       std::copy(global_image + i * g_width * 4,
                 global_image + (i + 1) * g_width * 4,
